use forward ref with as
