+++ packages/main-nav/src/index.ts
@@ -1,5 +1,5 @@
 export * from './MainNav';
-export * from './MainNavItem';
+export * from './MainNavLink';
 export * from './NavContainer';
 export * from './NavList';
 export * from './NavListItem';
